# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SRA Resource Management System API
  version: 1.0.0
  description: >
    Business-layer API for the SRA Resource Management System (SRA-RMS).
    Provides CRUD over Clients, Projects, Resources, and Allocations, plus
    dashboard and reporting endpoints. Authentication is delegated to
    Microsoft Active Directory (OAuth2 / OpenID Connect); authorization is
    role-based (Administrator, General, Report).
  contact:
    name: SRA
    email: umesh.kodippili@sra.com.au
  license:
    name: Proprietary - SRA
servers:
  - url: https://api.rms.sra.com.au/v1
    description: Production
  - url: https://api.staging.rms.sra.com.au/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Local development
security:
  - oauth2: []
...
paths:
...
  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Dashboard headline metrics
      operationId: getDashboardSummary
      description: Roles - Administrator, General.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Dashboard summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /dashboard/gantt:
    get:
      tags: [Dashboard]
      summary: Gantt data for projects and/or resources
      operationId: getGantt
      description: Roles - Administrator, General.
      parameters:
        - name: view
          in: query
          required: true
          schema: { type: string, enum: [projects, resources] }
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - name: department
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Gantt rows for the requested view and window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GanttResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------- Reports ----------------
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Microsoft Active Directory / Entra ID (OpenID Connect).
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            api://sra-rms/access_as_user: Access the SRA-RMS API as the signed-in user.
...
  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The authenticated user lacks the required role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    ProjectStatus:
      type: string
      enum: [planned, active, onHold, completed, cancelled]
...
    EffortUnit:
      type: string
      enum: [hoursPerWeek, percent]

    # ---- Client ----
...
    ProjectBase:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        code: { type: string, minLength: 1, maxLength: 50 }
        clientId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        budget: { type: number, format: double, minimum: 0 }
        remaining: { type: number, format: double, minimum: 0 }
        billable: { type: boolean, default: true }
        status: { $ref: '#/components/schemas/ProjectStatus' }
...
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          required: [id, name, code, clientId, startDate, endDate]
          properties:
            id: { type: string, format: uuid }
            clientName: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    AllocationBase:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        effort: { type: number, format: double, minimum: 0 }
        effortUnit: { $ref: '#/components/schemas/EffortUnit' }
        roleOnProject: { type: string }
        billable: { type: boolean }
...
    Allocation:
      allOf:
        - $ref: '#/components/schemas/AllocationBase'
        - type: object
          required: [id, projectId, resourceId, startDate, endDate, effort, effortUnit]
          properties:
            id: { type: string, format: uuid }
            projectId: { type: string, format: uuid }
            projectName: { type: string }
            resourceId: { type: string, format: uuid }
            resourceName: { type: string }
            warnings:
              type: array
              description: Non-blocking validation warnings, e.g. over-allocation.
              items: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    DashboardSummary:
      type: object
      properties:
        activeProjects: { type: integer }
        totalResources: { type: integer }
        averageUtilisation: { type: number, format: double, description: 'Fraction 0..1.' }
        overAllocatedResources: { type: integer }
        underAllocatedResources: { type: integer }
        budgetAtRisk: { type: number, format: double }
        upcomingProjectStarts:
          type: array
          items: { $ref: '#/components/schemas/Project' }
        upcomingRollOffs:
          type: array
          items: { $ref: '#/components/schemas/Allocation' }
    GanttResponse:
      type: object
      required: [view, from, to, rows]
      properties:
        view: { type: string, enum: [projects, resources] }
        from: { type: string, format: date }
        to: { type: string, format: date }
        rows:
          type: array
          items: { $ref: '#/components/schemas/GanttRow' }
    GanttRow:
      type: object
      required: [id, label, bars]
      properties:
        id: { type: string }
        label: { type: string }
        bars:
          type: array
          items: { $ref: '#/components/schemas/GanttBar' }
    GanttBar:
      type: object
      required: [start, end]
      properties:
        refId: { type: string, format: uuid }
        label: { type: string }
        start: { type: string, format: date }
        end: { type: string, format: date }
        effort: { type: number, format: double }
        overAllocated: { type: boolean }

    # ---- Reports ----
...